Start guide on using virtualenv for environment management
This will contain more details on how to setup the environment. Essentially,
this guide will help new team members setup a matching environment to the rest
of the team, thus easing transition.

In addition, it will ensure that everyone is using the same Python version and has
dependencies appropriately managed. Finally, using a virtual environment will
allow for self-contained projects.

As of now, the guide only contains instructions on how to setup the virtual
environment; it does not contain information on how to install and use different
Python versions (this will be added in the future).