Finish recipe selection page
Note: The styling is not done (yet). However, the selection page
has been built in React.

The server now has multiple new React classes, including:
	- Content: a general-purpose container to hold primary content
	- SectionHeaderButton: a header which has a label and a button with an action
	- SelectionList: a list of selection choices, as either checkboxes or
		radio buttons.
	- SelectionPage: layout for the index page to select recipes

In addition, the react-render.jsx file was renamed to _render.jsx, as
this allows the build tasks to append it at the end of react-components.jsx.
Otherwise, this causes a bug in which the server tries to render React classes
that occur later on in the file.