Modify build tasks to fix React ordering bug
As mentioned in cd82fc, there was an issue with how the build system
ordered the React classes. Now, this bug is fixed by prefixing files, which need
to be placed last, with an underscore.

In addition, the SASS build system was changed so that it requires a flag
to compile the Bootstrap components; this greatly speeds up build times
because the Bootstrap files rarely need to be compiled again.