Modify React modules (and building) to utilize browserify
Browserify is an npm module that allows us to use `require()` and `module.exports`
and build a bundle of dependencies instead of directly managing the script
loading. Thus, not only is it faster for development, it also allows for faster
client-side loading.

To make use of browserify, a good portion of the (React) codebase was rewritten.
First, the following React modules now use `require()` and `module.exports`:
	- Content
	- GlobalHeader
	- IngredientsPage
	- SectionHeaderButton
	- SelectionList
	- SelectionPage

Also, the "static/react" folder was moved to "static/js/react" to allow for simpler
`require()`s.

Unnecessary dependencies, which were included directly, were now removed:
	- tether.js
	- bootstrap.min.js

"_render.jsx" was moved to the primary "js" folder and renamed to "main.js";
this is the file that is compiled to "static/js/bundle.js".

The "package.json" file has been created to allow for easy installations of these
dependencies, using "npm install". Note: this also installs "browserify" globally.

Finally, the "bundle.js" was added to .gitignore so that it is not repeatedly
commited to the repository.