Add basic comments to React component classes
The individual rendering steps are not commented, however, as they are
mostly just JSX code.

The comments added (so far) just describe what the methods do; these comments
are not in great detail because the methods simply override methods provided
in React.Component. In addition, none of the render() methods take arguments.