Finish main Timer component and add demo to TimersPage component
The Timer component allows for stateful display of a decreasing time, which
changes with elapsed time. This is used directly by the TimersPage,
which will modify the timers accordingly.

The Timers use the `rc-progress` module's `Circle` component, which we can
modify the percentage filled for.

In addition, basic styling was added to these timers.