Disable "start" button on Timer page once started
This prevents confusion for the user, if (s)he attempts to press start again once
the app is already started.

In the future, this will most likely be replaced with a pause/forward control
panel, unless that panel is placed elsewhere.

To implement this, the SectionHeaderButton had to maintain a state (whether
the button is active or not). Then, the `changeButtonState()` modifies the state.