Display active and upcoming tasks, with correct timings
The mock data was edited to include the "end_time" field as well. Similar to our
LazyChef implementation, this time-based queue of tasks depends on a global
timer. However, instead of redrawing the timer with each new task, we utilize
React's management of state-based rendering to automatically swap tasks.

Currently, only the primary active task and top 3 upcoming tasks are rendered;
passive tasks and pause/skip features have not (yet) been added.

A small bug fix was implemented with filtering the references - we don't want to
progress every timer. Instead, we just need to progress the current primary and
all of the passive timers. This bug was fixed through a new filter regex.