Add keys for React components with dynamic amounts of children
When iteratively creating children for a React component, each child needs to
be given a unique key to identify it (on that page).

So, this was accomplished using `Math.random()`. The keys were assigned on
the following components, which were raising React warnings for not having
keys on the dynamic children:
	- IngredientsPage
	- SelectionList