Add Passive Task display of timers on a separate global timer hook
Instead of subscribing to the same hook, the passive tasks use a completely
separate timer so that pausing/stopping/skipping can function independently.

The passive timer rendering is offloaded to another component, because it has to
manage a dynamic list of what is currently running as a passive task.

Again, this uses a similar algorithm to the LazyChef rendering of passive timers,
except for the fact that the timers are decoupled.

One bug exists in this implementation (as of now) - the first passive task does
not display until the first active task completes, even though the passive timer
has the proper hooks at the proper times.