Add pause/skip functionality onto /timers page
To do so, the code for running the procedure per step had to be moved to a new
method in the TimerHooker class, which is the "progress" method - it is also used
internally by the timer to do its appropriate tasks.

Two methods had to be added to the TimersPage component, which handle the
"pause" and "skip" events, respectively. In addition, the Ionicons style sheet was
added to the head HTML so that the pause/forward icons could be used. Similarly,
a CSS class was created for these icons.