Fix passive timer startup bug
As mentioned in e1b4da54, there was a bug with how passive timers started.
For some reason, they would only launch after the first active task was completed.

As mentioned in the comment in the queueTasks method in PassiveTaskList.jsx,
this was an issue due to how React detects state changes - the "data" object
is stored as a reference and so, modifying the object does not trigger a state change.
As a solution, we now use a counter that updates when new items are added to
the data object, which triggers the appropriate state change and as a result,
React renders the new timers.

As a minor fix, the pause/skip buttons no longer trigger their callbacks when
they are disabled.