Display description in addition to task name under primary active timer
This required slightly modifying the Timer component. The largest modification,
however, came from the positioning the current time remaining in the center
of the text. To position correctly, a bit of manipulation of the "left" CSS property
was required, but it does render in the correct position now.

In addition, the signature of Timer.reset was changed (to include the description),
and so uses of this method had to be appropriately modified.