Style improvements across all pages
The main styling of the pages has been started. This includes ensuring that the
website works equally well on both mobile and the desktop.

A key aspect to this was having properly-sized timers so that they do not
overflow and are still large enough to be useful. This was capped at 5em, which
met both criteria.

Similarly, the primary active timer is capped at a size of 10em.

On another note, the ingredients page now pads the ingredients listed.