Only display top 3 passive tasks with the shortest duration remaining
This is a much simpler algorithm than the previous used, and is also much more
logical. As a result of the simplicity, it was moved directly into TimersPage.jsx.
Thus, PassiveTaskList.jsx was deleted because we no longer depend on it.

In essence, the algorithm marks a passive task as available to display when it
hits the start time, and marks it as no longer available when it hits the end time.
Then, all of the available tasks are filtered and sorted by their end time.
The three with the lowest end times are displayed.

For passive tasks, a timer display cannot be fully removed, because it may be
utilized later on. So, the Timer.remove method now has a flag for whether
to remove permanently (thus never rendering again) or temporarily (renders
again when needed).