Display selected recipes on /timer page
To accomplish this, a small bug had to be fixed: the recipes were not being
passed to the Server.get_timers router. This bug was due to the form itself being
overwritten by React. To fix the bug, the React element was moved to a nested div,
and so the form data now remains intact (and is thus passed to the router).

The recipes list itself is a collapsible panel, which contains rows of the given
recipes. The list is housed as an instance of the CollapsibleList component.