Add global timer stop functionality: pause both active/passive timers
The new stop button pauses both the active and passive timers. When stopped,
the timers can be (simultaneously) restarted by pressing the play button, which
replaces the original pause button.

The initial pause functionality still remains the same. However, it is important
to note that the app cannot be stopped AND paused at the same time - only
one or the other is possible and the appropriate buttons/graphics ensure this.