Fixed issue with active tasks starting with one second removed
Because active tasks ended at task.end_time + 1, this meant that the next task
was not displayed until task.start_time + 1. Thus, this meant that the task timer
would always start with 1 second already removed.

By removing the erroneous +1 for the end time, the bug was fixed. Similarly,
the code in _skipCurrentTask() was modified to match the correct timing scheme.

This fix was also applied to the passive tasks, in which the bug also persisted.