Optimize stateless components as functions instead of ES6 classes
By doing so, React does not store or attempt to manage state, and this makes
those parts of the application faster to render and manage. In essence, these
components now act as templates rather than stateful components.

The components modified to be stateless are:
	- Content
	- GlobalHeader
	- IngredientsPage
	- SelectionList
	- SelectionPage

This commit closes #3, because the stateless components have now been optimized.