Create custom checkbutton/radio React components for simpler styling
Although using pure CSS is a more elegant solution, as shown here:
http://stackoverflow.com/a/27779075/1730261, it is arguably more complex.
In addition, CSS alone did not play well with switching icons when the checkbutton
was checked.

Instead, a React component was created for "SelectionInput"s, which are essentially
checkbuttons or radioboxes. Through the use of state changes and a hidden input,
it maintains the same functionality as the CSS variant.
Minimal CSS was utilized to handle hiding the literal checkbox and styling
the icons.

Appropriately, SelectionList now uses SelectionInput rather than ReactBootstrap.Input.