Start configuration file templates and automatic packaging
In essence, running "python package.py" will create a new file, "configure.py",
that can be run on the server to automatically configure it. It has all of the
data for the server (including configuration files) bundled within.

The services currently configured, along with their respective files:
	- NginX: nginx.conf
	- Upstart: upstart-app.conf

In addition, the "init" file installs the required packages (on the server)
and also orchestrates the entire setup. So, "configure.py" simply outputs
its stored data to various configuration files, and then calls "init".

Finally, "configure.py" was added to the gitignore because it changes frequently
and can be built on-demand.