Set up uWSGI configuration for accepting proxied requests from NginX
To utilize uWSGI, the Flask Server needs to be directly accessible, without the
Application wrapper. So, the Application.getServer() static method was created
to accomplish creating a new application and obtaining its server instance.

In addition, the uWSGI configuration file (uwsgi.ini) was created to facilitate
running uWSGI from the directory. This uses the "wsgi" module to run, and so
wsgi.py had to created. In addition, the Upstart script had to be modified so
that it supplies the virtualenv path to uWSGI.

Finally, some directory configuration was added to config.py and these changes
were propagated throughout the configuration-file templates.