Added Parse analytics IDs to elements, with support for "ID inheritance"
Because React props can be propagated throughout a component's children, the
given Parse ID for a container element will be passed onto its children. However,
only one specific child (generally, a button or action item) will make use of the
Parse ID and set it as its own.

The one exception to this behavior is a container whose children should all have
a unique ID, such as a container of selections (namely, SelectionList). Thus, the ID
for the container is used as a "base" for the children IDs, which are generated, in
this case, based on the value of each selection.

For all components that make use of the Parse ID, however, passing it is optional.

Another minor change included renaming the RecipeList to CollapsibleList in
CollapsibleList.jsx; the class was simply named incorrectly.