Configure Vagrant to install and run Ansible playbooks on VM
Although Ansible generally runs from the host and executes code on the guest,
we cannot run Ansible from our environment directly, because we use Python 3.5.0,
and Ansible requires 2.6-2.7. So, Ansible runs on the VM directly using the
"Ansible Local" feature.

Currently, installation with pip is not working, and so virtualenv is installed
via apt-get instead. Finally, the Upstart configuration was modified to match
the appropriate directory on the server. However, the "uwsgi" call in this
configuration is not working yet.