Use a Shell provisioner to install and run Ansible through Vagrant automatically
Fixes the issue introduced in f30101, because now we do not rely on Vagrant's
"ansible_local" provisioner. Instead, a shell script (ansible.sh) installs and
runs Ansible to provision the server. In the future, once Vagrant's issue
with "ansible_local" is fixed, we will switch over to their provisioning:
http://stackoverflow.com/questions/35299304/automatically-installing-and-running-ansible-local-via-vagrant/35304194#35304194.

In addition, the box version is restrited so that incompatibilities with modules
do not occur; we want easy consistency and restricting the version ensures that.