Move configuration to configuration file and allow Ansible to include configuration
By using this level of modularity, both Ansible and the app can share the same
configuration variables. In addition, the configuration is now split into three
groups, in the same location (config.yml):
	- default
	- production
	- development

The "default" group is the same as production; it is all of the default settings.
However, some of these settings are overridden in the "development" group,
for when we are running a development environment. As such, the config module
makes sure to recursively merge the default group with the current environment.

In addition, the Ansible playbook has more tasks; namely, creating various
directories for the server and setting up the firewall.