Enable Ansible to use different environment variables
To do this, the configuration had to be moved under one single key in config.yml,
named "config". This allows Ansible to dynamically find the correct dictionaries
to merge. For Ansible to actually see the current environment, it must be passed
as an argument to "vagrant up", which is then passed to ansible.sh.

Configuration variables in Ansible are stored as part of the "config" dictionary,
instead of the "default", which was just the name of the environment.

Finally, old initialization snippets/templates were deleted.

Other changes include:
	- enforced Vagrant version in Vagrantfile
	- moved template files to templates directory