Configure SSH settings on server via Ansible
This uploads your public key (whose path is set by the SSH_KEY environment variable
or the default path, ~/.ssh/id_rsa.pub) to the server, and then allows that key to
SSH into the server, by appending it to the user's ~/.ssh/authorized_keys file.

Finally, the user's name was changed to "admin" for simplicity.