Turn bin_pack into subpackage of new lib package
The "lib" package contains all of the packages developed for the app.
"bin_pack", which contains the main algorithm, is stored in the "core"
subpackage.

Note: most of the testing code was moved to the lib/core/tests folder, but
will not work as of now. Eventually, the tests should be moved to a dedicated
unit tests folder, which will also test any (and all) other code in the project.

Overall, this should allow for simpler organization and management of
various packages.

Finally, modules and classes were renamed to follow Python conventions.