Revert imports to package-level imports and switched to raising errors
Instead of printing an error message and returning None (or worse, letting
the error slip by), the DatabaseManager now raises errors for when an
operation fails.

In addition, the database.py module was cleaned up. Instead of forcing an
initialization, the DatabaseClient checks if the file exists or not prior to
creating the connection - if it does not, then it initializes the database.