Allow server to fetch recipes from database
Instead of fetching everytime, however, we used a CachedValue which essentially
caches the fetched recipes for a set time (in this case, a minute) and then
refreshes the cache when necessary. Although primitive, it works well for
minimizing unnecessary database calls.

In addition, a configuration option for the database file was added.