Integrate database with main app server
Two new caches were added to `lib.core.cache`, and provide functionality for
a dict-type cache, except with future portability.

In particular, the FixedSizeCache is used to prevent too many meals (where a meal
is defined as a set of recipes) from being cached at once - each meal takes a
relatively large amount of memory to store, but also a lengthy time to compute,
so a fixed cache seems like an ideal solution.

However, an error occurs with attempting to access the sqlite3 database from
a request - the connection cannot be used in another thread (i.e the request).
See #4 for more details.

Other changes include:
	- catching exceptions when sequence is not in database instead of explicitly
	checking
	- closing database mechanism