Fix multiple-thread database accesses with sqlite3
The bug introduced in #4 is now fixed, but this fix is hopefully temporary. It is
most likely not safe to disable checking if the database is being accessed
from the same thread, so it is worthwhile to look for a more permanent solution.

Closes #4.

In addition, all cursor connections are closed once they are done being used.