Redirect GET requests from /timers and /ingredients to /index
This prevents the "Method Not Allowed" HTTP response when accidentally
sending a GET request to the /timers page as opposed to a POST request.
In the future, it may be worthwhile redirecting to using POST if a previous
POST request was made to the same endpoint.

Also, response handler names were changed to more appropriately reflect
what they respond to.