Install pyenv and appropriate Python version through Ansible
To do so, the environment has to be set before installing packages - this allows
the `pyenv` command to be recognized. This led to separating the environment
variables from the standard setup through "tasks/env.yml".

In addition, the git command was changed so that "pyenv" is cloned as the
admin user instead of root.