Start developing recipe input page
The page is served on /new, but may be moved to a secondary endpoint later on
(i.e /recipes/new).

A few new React components were introduced for this page, namely the NewRecipePage,
InputRow, and StepInput.

In addition, a few (code) stylistic changes were made to SelectionList,
SectionHeaderButton, and FullRow.

This commit begins the work on #8.