Add save feature on recipe input page
In addition, support for the min_wait and max_wait times was added, which
is a required part of a TaskNode (but neglected in the prior commit).

To actually transmit the task data, it is first transformed into JSON on the
client-side, and then parsed on the server. This seems a bit excessive and
counter-intuitive, but it was the simplest way to transfer the array, containing
objects, of unknown size to the server. In the future, we may end up switching
completely to JSON and use the `POST /new` endpoint as more of an API
endpoint rather than a regular form endpoint.

On another note, random "todo" comments were removed from the Bootstrap
SASS files, so that `git grep todo` and `git grep TODO` only search our own code
for TODO items.