Add recipe-specific router (as a blueprint) and abstract it away from main server
Not only does this provide modularity, but it also allows for easier maintenance
later on, such has switching to a different endpoint.

As such, the recipe input routes were migrated to the new blueprint, which is
where they belong.

To help support modularity, a new package was created in the "lib" package:
`lib.blueprints`, which contains the various blueprints for the application.