Start Recipes viewing page which will allow for database updates
Currently, only the layout of the routers and components have been created,
but nothing is actually displayed. Soon, however, the appropriate views and
controls will be added so that recipes can be viewed, modified, or deleted.

In addition, a new classmethod was added to the Router wrapper so that a
dictionary of values can be shared globally among all routers, as opposed to
manually entering each value separately.