Create an EditableSpan that allows for editing of values based on state
Through this, basic recipe editing functionalities were implemented. For example,
steps can be edited for their name, description, and numeric values. In addition,
new steps can be added to a specific recipe and saved to the database.

However, deleting steps and adding them in between other tasks is not yet supported.