Refactor EditRecipePage and NewRecipePage to consolidate shared code into NewStep
The NewStep component manages adding components and resetting the input once done.
To add the data to the relevant arrays in EditRecipePage and NewRecipePage components,
the NewStep takes a callback which is passed the new task.

This improves modularity and allows for simpler code changes later on.

In addition, the /recipe/edit/ POST request now redirects to /recipe/ as opposed to
the same page.